Guard against null temporary files (?). Don't call getTestClass() because it loads it using caller's class loader.

+++ junit4-ant/src/main/java/com/carrotsearch/ant/tasks/junit4/JUnit4.java
@@ -862,7 +862,9 @@ public class JUnit4 extends Task {
     if (!leaveTemporary) {
       for (File f : temporaryFiles) {
         try {
-          Files.deleteRecursively(f);
+          if (f != null) {
+            Files.deleteRecursively(f);
+          }
         } catch (IOException e) {
           log("Could not remove temporary path: " + f.getAbsolutePath(), Project.MSG_WARN);
         }
